[java] Refactor security rules
