+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/security/InsecureCryptoIvRule.java
@@ -4,16 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.rule.security;
 
-import net.sourceforge.pmd.lang.java.ast.ASTArgumentList;
-import net.sourceforge.pmd.lang.java.ast.ASTArrayAllocation;
-import net.sourceforge.pmd.lang.java.ast.ASTArrayInitializer;
 import net.sourceforge.pmd.lang.java.ast.ASTConstructorCall;
-import net.sourceforge.pmd.lang.java.ast.ASTExpression;
-import net.sourceforge.pmd.lang.java.ast.ASTStringLiteral;
-import net.sourceforge.pmd.lang.java.ast.ASTVariableAccess;
-import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
-import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
-import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
 
 /**
  * Finds hardcoded static Initialization Vectors vectors used with cryptographic
@@ -31,51 +22,9 @@ import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
  * @since 6.3.0
  *
  */
-public class InsecureCryptoIvRule extends AbstractJavaRulechainRule {
-
-    private static final Class<?> IV_PARAMETER_SPEC = javax.crypto.spec.IvParameterSpec.class;
+public class InsecureCryptoIvRule extends HardCodedConstructorArgsBaseRule {
 
     public InsecureCryptoIvRule() {
-        super(ASTConstructorCall.class);
-    }
-
-    @Override
-    public Object visit(ASTConstructorCall node, Object data) {
-        if (TypeTestUtil.isA(IV_PARAMETER_SPEC, node)) {
-            ASTArgumentList arguments = node.getArguments();
-            if (arguments.size() > 0) {
-                validateProperIv(data, arguments.get(0));
-            }
-        }
-        return data;
-    }
-
-    private void validateProperIv(Object data, ASTExpression firstArgumentExpression) {
-        if (firstArgumentExpression == null) {
-            return;
-        }
-
-        // named variable
-        if (firstArgumentExpression instanceof ASTVariableAccess) {
-            ASTVariableAccess varAccess = (ASTVariableAccess) firstArgumentExpression;
-            if (varAccess.getSignature() != null && varAccess.getSignature().getSymbol() != null) {
-                ASTVariableDeclaratorId varDecl = varAccess.getSignature().getSymbol().tryGetNode();
-                validateProperIv(data, varDecl.getInitializer());
-            }
-        }
-
-        // hard coded array
-        if (firstArgumentExpression instanceof ASTArrayAllocation) {
-            ASTArrayInitializer arrayInit = ((ASTArrayAllocation) firstArgumentExpression).getArrayInitializer();
-            if (arrayInit != null) {
-                addViolation(data, arrayInit);
-            }
-        }
-
-        // string literal
-        ASTStringLiteral literal = firstArgumentExpression.descendants(ASTStringLiteral.class).first();
-        if (literal != null) {
-            addViolation(data, literal);
-        }
+        super(ASTConstructorCall.class, javax.crypto.spec.IvParameterSpec.class);
     }
 }
